rename functions and add deletion of uvs an normals shared arrays, move the code in dedicated file
