+++ libs/common/common_utils.h
@@ -33,6 +33,16 @@
 
 PXR_NAMESPACE_OPEN_SCOPE
 
+#if ARNOLD_VERSION < 70307
+constexpr auto AiArrayMapConst = AiArrayMap;
+constexpr auto AiArrayUnmapConst = AiArrayUnmap;
+#endif
+
+
+#if ARNOLD_VERSION_NUM < 70100
+constexpr auto AiArnoldIsActive = AiUniverseIsActive;
+#endif
+
 // Older versions of Arnold had imagers defined as drivers
 #if ARNOLD_VERSION_NUM < 70301
 static const int AI_NODE_IMAGER = AI_NODE_DRIVER;
