+++ libs/render_delegate/render_delegate.cpp
@@ -469,11 +469,8 @@ HdArnoldRenderDelegate::HdArnoldRenderDelegate(bool isBatch, const TfToken &cont
     _id = SdfPath(TfToken(TfStringPrintf("/HdArnoldRenderDelegate_%p", this)));
     // We first need to check if arnold has already been initialized.
     // If not, we need to call AiBegin, and the destructor on we'll call AiEnd
-#if ARNOLD_VERSION_NUM >= 70100
     _isArnoldActive = AiArnoldIsActive();
-#else
-    _isArnoldActive = AiUniverseIsActive();
-#endif
+
     if (_isBatch) {
 #if ARNOLD_VERSION_NUM >= 70104
         // Ensure that the ADP dialog box will not pop up and hang the application
